# Conflicts:
#	Diablo2/Game/data.zip

+++ Diablo2/Motor2D/j1Pathfinding.h
@@ -88,19 +88,19 @@ struct PathNode
 {
 	// Convenient constructors
 	PathNode();
-	PathNode(int g, int h, const iPoint& pos, PathNode* parent);
+	PathNode(float g, float h, const iPoint& pos, PathNode* parent);
 	PathNode(PathNode& node);
 
 	// Fills a list (PathList) of all valid adjacent pathnodes
 	uint FindWalkableAdjacents(PathList& list_to_fill);
 	// Calculates this tile score
-	int Score() const;
+	float Score() const;
 	// Calculate the F for a specific destination tile
-	int CalculateF(const iPoint& destination);
+	float CalculateF(const iPoint& destination);
 
 	// -----------
-	int g;
-	int h;
+	float g;
+	float h;
 	iPoint pos;
 	PathNode* parent;
 };
